DTO cleanup

+++ src/main/java/io/contek/invoker/binancespot/api/common/_BookTicker.java
@@ -1,13 +1,14 @@
 package io.contek.invoker.binancespot.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
+import java.math.BigDecimal;
 
 @NotThreadSafe
 public class _BookTicker {
 
   public String symbol;
-  public Double bidPrice;
-  public Double bidQty;
-  public Double askPrice;
-  public Double askQty;
+  public BigDecimal bidPrice;
+  public BigDecimal bidQty;
+  public BigDecimal askPrice;
+  public BigDecimal askQty;
 }
